Update format, move hooks
