+++ packages/web/src/core/stores/index.ts
@@ -1,4 +1,4 @@
-import { useDeviceContext } from "@app/core/stores/utils/useDeviceContext";
+import { useDeviceContext } from "@app/core/hooks/useDeviceContext";
 import { type Device, useDeviceStore } from "@core/stores/deviceStore";
 import { type MessageStore, useMessageStore } from "@core/stores/messageStore";
 import { type NodeDB, useNodeDBStore } from "@core/stores/nodeDBStore";
@@ -7,7 +7,7 @@ export {
   CurrentDeviceContext,
   type DeviceContext,
   useDeviceContext,
-} from "@app/core/stores/utils/useDeviceContext";
+} from "@app/core/hooks/useDeviceContext";
 export { useAppStore } from "@core/stores/appStore";
 export {
   type Device,
@@ -28,7 +28,6 @@ export {
   SidebarProvider,
   useSidebar, // TODO: Bring hook into this file
 } from "@core/stores/sidebarStore";
-export { useNewNodeNum } from "@core/stores/utils/useNewNodeNum";
 
 // Define hooks to access the stores
 export const useNodeDB = (): NodeDB => {
